* Refactor Ferrum module * Split methods into Utils
